Looking at lsof(1) output, it seems that unmounting of /old_root/dev fails because of the kernel daemons keeping /dev/console open, which doesn't seem very proper to me. No matter, the right way to move a filesystem IMO is with mount --move, not by umounting-then-mounting it. The only drawback is that MS_MOVE has to be #defined [to 8192].
Mounting initrd's /bin as /usr/local/bin was quite dumb
of me I must admit. It's much easier, if needed, to leave
/old_root and put /old_root/bin in $PATH. That's not FHS
compliant, so s/old_root/local/g * if you wish.

2nd_stage_small.tar.gz is created with tar | gzip -9
because it's size is very near the limit (that is, with
mine linked as shared, which currently gives a sig11).
Using .bz2 will be better, if approved by Rene/Clifford.

2nd_stage_small (re)boots/shutdowns with no errors, mount
lists the filesystems mounted properly and /old_root can
be unmounted, so 2nd_stage full should not display errors
as well. Now if there was anybody willing to fix mine...


git-svn-id: https://svn.exactcode.de/t2/trunk@1599 c5f82cb5-29bc-0310-9cd0-bff59a50e3bc