Create pciutils.desc
the pci.ids.bz2 package changes regularly so having both a set check code and dynamic name should not occur otherwise the .desc file would have to be rebuilt daily (and sometimes apparently multiple times a day) just to keep it to match unless you intend storing them in your packages system and that is not advised as good by the group maintaining the pciutils package. this is my workaround for that, how you want to do it to match your T2 system is up to you.